# playwright/support/zod/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Restful Booker Platform API",
    "version": "1.0.0",
    "description": "API documentation for the Restful Booker Platform"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/rooms": {
      "get": {
        "summary": "Get all rooms",
        "tags": [
          "Rooms"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of rooms",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "number": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string",
                            "enum": [
                              "STANDARD",
                              "DELUXE",
                              "SUITE"
                            ]
                          },
                          "price": {
                            "type": "number"
                          },
                          "capacity": {
                            "type": "number"
                          },
                          "amenities": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "AVAILABLE",
                              "OCCUPIED",
                              "MAINTENANCE"
                            ]
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create new room (admin only)",
        "tags": [
          "Rooms"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "number",
                  "type",
                  "price",
                  "capacity"
                ],
                "properties": {
                  "number": {
                    "type": "string",
                    "example": "101"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "STANDARD",
                      "DELUXE",
                      "SUITE"
                    ],
                    "example": "STANDARD"
                  },
                  "price": {
                    "type": "number",
                    "example": 100
                  },
                  "capacity": {
                    "type": "number",
                    "example": 2
                  },
                  "amenities": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "WiFi",
                      "TV"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Room created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "number": {
                          "type": "string"
                        },
                        "type": {
                          "type": "string"
                        },
                        "price": {
                          "type": "number"
                        },
                        "capacity": {
                          "type": "number"
                        },
                        "amenities": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "status": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          }
        }
      }
    },
    "/api/rooms/{id}": {
      "get": {
        "summary": "Get room by ID",
        "tags": [
          "Rooms"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Room ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Room details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "number": {
                          "type": "string"
                        },
                        "type": {
                          "type": "string",
                          "enum": [
                            "STANDARD",
                            "DELUXE",
                            "SUITE"
                          ]
                        },
                        "price": {
                          "type": "number"
                        },
                        "capacity": {
                          "type": "number"
                        },
                        "amenities": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "AVAILABLE",
                            "OCCUPIED",
                            "MAINTENANCE"
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Room not found"
          }
        }
      },
      "put": {
        "summary": "Update room (admin only)",
        "tags": [
          "Rooms"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Room ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "number": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "STANDARD",
                      "DELUXE",
                      "SUITE"
                    ]
                  },
                  "price": {
                    "type": "number"
                  },
                  "capacity": {
                    "type": "number"
                  },
                  "amenities": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "AVAILABLE",
                      "OCCUPIED",
                      "MAINTENANCE"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Room updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "number": {
                          "type": "string"
                        },
                        "type": {
                          "type": "string"
                        },
                        "price": {
                          "type": "number"
                        },
                        "capacity": {
                          "type": "number"
                        },
                        "amenities": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "status": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "404": {
            "description": "Room not found"
          }
        }
      },
      "delete": {
        "summary": "Delete room (admin only)",
        "tags": [
          "Rooms"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Room ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Room deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "404": {
            "description": "Room not found"
          }
        }
      }
    }
  },
...
}